Refactor RoleMap methods to remove utils dependency.
